Allow INotifyPropertyChanged and use weak references
Prior to this change, the event handling tied the lifecycles of the two objects. While this was probably not a big problem since the command usually lives on the view-model it subscribes to, this solution is more general. Now the command can also subscribe more generally to any object implementing INotifyPropertyChanged