Improve use of ILogger
Prior to this change, the ILogger was created by calling a static factory.

This change allows the process of creating the logger to be modified to allow for easy creation by an IoC container. The static manager class is still in case as a factory the IoC container can utilize. The manager is also able to detect in what environment the framework is running and will create an appropriate logger type.