Fix Dispatcher use
Prior to this change, I messed up and used the current dispatcher instead of the main thread one.

This change uses the correct dispatcher and also avoids the necessity to initialize it on the WPF side. A future switch to using the Noesis Dispatcher is possible once it implements the InvokeAsync call.