Use builtin IoC container in Bootstrapper base
Prior to this change, the base implementation did not have any IoC support. Also AssemblySource was adding the executing assembly on its own. This should be handled by the bootstrapper method specified for this purpose.

This change lets the bootstrapper use the bult-in MiniIoC container by default and moves the adding of assemblies to the SelectAssemblies method.